refactor: introduce markdown-document module
