+++ AngularStandalone/src/app/app-routing.module.ts
@@ -1,9 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { DocumentComponent } from './document/document.component';
+
 import { HomeComponent } from './home/home.component';
 import { NetCoreApiComponent } from './net-core-api/net-core-api.component';
-import { SearchComponent } from './search/search.component';
+import { SearchComponent } from './markdown-document/search/search.component';
+import { DisplayComponent } from './markdown-document/display/display.component';
 
 const routes: Routes = [
   { path: '', component: HomeComponent },
@@ -12,7 +13,7 @@ const routes: Routes = [
     path: 'doc/:ref', children: [
       {
         path: '**',
-        component: DocumentComponent,
+        component: DisplayComponent,
       },
     ],
   },
