+++ AngularStandalone/src/app/markdown-document/display/display.component.html
@@ -1,0 +1,2 @@
+<app-document></app-document>
+
