+++ AngularStandalone/src/app/shared/document/document.component.html
