+++ AngularStandalone/src/app/shared/document/document.component.scss
